OF-2611: S2S Unit test: clean up TODOs, add spec references
This adds references to RFCs in the ExpectedOutcome calculation, and removes some of the TODO statements in that class.

Most of the TODOs are 'resolved' by adding a 'strictCertificateValidation' setting in the calculation. This is used to choose between the multiple possible outcomes that were in the 'TODO'.

Note taht the strictCertificateValidation setting is hardcoded in all tests. The current implementation makes it hard to configure different values for the initiating and receiving entities. Also, they would add to an already long list of tests.