Remove duplication of connection configuration
The Connection interface defines methods to read the configuration of TLS and compression policies, even though these are also defined by the ConnectionConfiguration instance that is used to create the connection.

It is undesirable to have the configuration of a connection be defined in various places, or be modified after the original connection has been applied.

This commit removes the duplication, and ensures that connection configuration is applied as soon as the instance is created.

As a side-effect, this solves an issue with the new Netty code, that never explicitly sets the tlsPolicy on the connection.

The single functional aspect of the separation of tlsPolicy between connection and configuration (prior to this change) was the following: the state of the connection-tlsPolicy was used to implicitly define if a session was initialized (this was used to close a connection that was sending unencrypted data, when its configuration required encryption). This commit replaces that implicit defintion by a new, explicit 'isInitialized` method on the Connection interface.