OF-2559: Netty SslContext should have Openfire's TrustManager
Key and Trust material is used when creating an SSL context. Openfire provides custom implementations, that tie into its own keystores (the 'identity' and 'key' stores).

In one of the new SslContext creation methods that are specific to Netty, Openfire's custom truststore was never added to the context that was being created. As a result, trust material that is configured to be used by Openfire was not. This problem might not always be apparent, as I believe that without explicit configuration, default Java trust material is used. There will be much overlap between generally accepted trust material in Openfire's keystores, and the defaults in Java.

This commit adds Openfire's specific trust store to the SSL context that is going to be used by Netty.