OF-2638: Firewall documentation improvements
The firewall section of the install guide should make it clear that even though the admin console _can_ be exposed to the general internet, it should not be.

By highlighting port numbers and using paragraphs per subject, the text becomes easier to consume.

The old text discarded the DirectTLS ports. At some point in time, it was believed that these would be phased out. As that's not the case, they should be mentioned in the guide.