OF-2639: No longer require authzid for S2S SASL EXTERNAL
In the old days, the 'from' attribute on a stream element was not present. This made providing an authorization identity during the SASL handshake a requirement (otherwise, you'd not know who you'd be talking to).

Now that the 'from' attribute can be relied on to provide the identity of the peer, it does not need to be included in the SASL EXTERNAL handshake any longer.

This commit drops this requirement, but adds a new property (`xmpp.auth.sasl.external.server.require-authzid`) that can re-introduce it.