OF-2647: Restore JiveGlobals.getHomeDirectory(String)
A previous commit for OF-2647 replaced the argument to the getter and setter of `JiveGlobals.setHomeDirectory()` from a String to a Path.

This leads to compatibility issues in various plugins:
- https://github.com/igniterealtime/openfire-bookmarks-plugin/issues/20
- https://github.com/igniterealtime/openfire-certificateManager-plugin/issues/19
- https://github.com/igniterealtime/openfire-monitoring-plugin/issues/362

In hindsight, the signature change was to much of a change. This commit 'undoes' that change, by:
- introducing `get/setHomePath(Path)` (instead of replacing `getHomeDirectory`)
- restoring `get/setHomeDirectory(String)` and marking it as being deprecated

This should restore backwards compatilibity, and give plugin implementors time to migrate from the old to the new API.