OF-2648: Restore restoration of stream namespaces
In OF-2556, a workaround was provided to 'restore' namespace declarations on a stream element that had been parsed during the setup of the session.

This workaround injected a prefix-less 'stream' element. Unless the original stream defined a default namespace, this will not properly be parsed by an XML parser.

In this commit, the prefix 'stream' is used for the injected element (which is a de-facto standard in XMPP), and, for good measure, the namespace for that prefix is added, if it wasn't already recorded before.