OF-2658: dynamically add/remove Netty ChannelHandlers to SocketAcceptors
This commit introduces an API that lets plugins add providers of ChannelHandler instances, that are to be added to channels of a SocketAcceptor.

With this API, plugins can modify the behavior in Netty channel pipelines. One application for this is the XmlDebugger plugin, which adds a ChannelHandler that logs data that is passing through the pipeline.