OF-2659: Do not use received namespaces on outbound stream
Openfire keeps track of namespace declarations on `stream` elements that it receives. These are intended to be used when parsing child elements, which are provided later / separately of their original stream root element. See OF-2556 for details.

These recorded namespaces should only be used for parsing data that's received in the same stream. Erroneously, Openfire also used them in an outbound stream definition. This introduced duplication of certain elements, causing remote servers to fail XML parsing.