Fix: Correctly clean up OutgoingServerSession
On cleanup of old S2S sessions, the `OutgoingServerSessionListener#onConnectionClose` is called for the outbound connection, at this time the routing table contains a single route, but the route is not removed because the session passed to `OutgoingServerSessionListener#onConnectionClose` as the "`handback`" parameter has an empty `outgoingDomainPairs` field.

So when `OutgoingServerSessionListener#onConnectionClose` tries to `removeServerRoute(domainPair)` by iterating over `session.getOutgoingDomainPairs()` there is no `domainPair` to remove.

Adding the Outgoing Domain Pair to the new session on `transferConnectionToNewSession` allows the route to be cleaned up correctly.