CI: skip unit tests when performing database-upgrade test
The Openfire unit tests are executed as part of other CI jobs. By running them as part of the build of Openfire that's used for the database upgrades, those database upgrades might report failure when in reality the unit tests fail, and will add considerable amount of runtime to the job execution.