Stabilize LocalOutgoingServerSessionTest outcome
A race condition can cause LocalOutgoingServerSessionTest to occasionally fail. This seems to affect the Github based runners more than my local machine.

The issue is caused by a NullPointerException that is thrown immediately after a S2S connection has been set up. This causes the connection to close at the same time that the connection state is evaluated.

This commit prevents the NPE by adding a mock for the routing table to the test fixture.