OF-2665: Ensure correct cache config for SessionManager caches
Various caches used by SessionManager are expected to be configured as unlimited, never expire. However, they had no explicit configuration, therefor had the default size limitation and expiry. This leads to cache inconsistency issues as documented in OF-2665

For two out of three caches, it appears that the names of the caches have at some point been modified, without that same change having been applied to the cache configuration. The third cache is new, and never had cache configuration added.

This commit fixes that, be updating/adding cache configurion.

Special care needs to be taken with regards to a second set of cache configuration, that lives in the Hazelcast plugin. The names used in Openfire and there should be equal!