OF-2682: Replace static field on Connection with abstract class
The 'namespaces' field defined on the Connection interface was expected to contain namespace delarations specific to the Connection instance.

However, fields defined on a Java interface are static, meaning that their content is shared with all instances.

To ensure that each instance gets its own data, the field has been moved into a new abstract class. Existing implementations are changed to inherit from this new class.

There might be potential to move more shared functionality from implementations to this new shared class. That has not been done in this commit.