OF-2697: Improve concurrent setup of S2S connections
By replacing the queue in the thread pool used to setup server-to-server connections, it is far more likely that connections will be created concurrently.

The old queue size property served two purposes. As the thread pool queue now no longer has a configurable capacity, the property description is modified to reflect its other usage: that of the maximum size of the queue of stanzas waiting to be delivered.