OF-2708: Ensure that bare JIDs are used with groups
A Group is ment to be a collection of bare JIDs, not full JIDs. This is illustrated by the default provider, which stores only the node-part of a JID for local users.

This commit ensures that any JID that is used when interacting with a group is used in its 'bare' representation. This prevents an issue where an invocation of remove(fullJid) would not be successfull when the group did contain the corresponding bare JID of that entity.