OF-2668 / OF-2723: Allow plugins to be compiled against Openfire 4.8
In this commit, a second artifact is produced from the xmppserver module. Apart from the prexisting 'jar' artifact, there now is a 'taglib' artifact.

This taglib artifact is a jar file (that needs to go into our maven repository), that holds the Admin Console taglibs, in a spec-compliant directory structure.

The plugin's parent pom-file is modified to be able to make use of this new artifact.

With this, compilation of a plugin that defins the plugin parent pom as its parent succeeds (although an unhealty amount of errors is logged). The resulting admin console pages appear functional, at least in Openfire 4.8.