OF-2728: Phase out Rome
The admin console shows a bit of RSS data. The RSS functionality that we use is very limited. Instead of using a full-fledged RSS parser (that historically has triggered a lot of static analysis warnings), we might as well roll our own implementation, based on all of the XML parsing implementation that we already have for XMPP parsing.