OF-2733: Sync Openfire's truststore with Mozilla's shipped CAs
Before we cut a new release, let’s make sure that we update the content of our truststore with the latest CAs. We traditionally use the ones from Mozilla.

Dan created a handy script for this task. You can find it in the build directory.

This commit replaces the truststore with one that's created with that script.