OF-880: No Such User detection to evaluate pre-authenticated session
A pre-authenticed session should not be consided a 'no such user' case.

Given the implementation, it's hard to identify a pre-authenticated session based on a (bare JID) 'to' address. The API is designed to be used by 'from' addresses.

Pre-authenticated sessions are unlikely to be the subject of a query, unless it is the client itself issuing the query. Aioxmpp's tests have been observed to perform disco/info requests prior to authentication. Given this, we can use the 'from' address (full JID) if its bare JID matches the addressee for the 'no such user' detection.