OF-2738: Include SNI when setting up outbound server-to-server connection
When setting up an outbound server-to-server connection, Openfire should send along an indication to what domain name it is attempting to connect. This will help the recipient server to process the connection for the correct vhost when that server is hosting multipe XMPP domains on the same server (ejabberd allows for this, for example).