CI: speed up artifact sharing
In some (but not all - this might be infrastructure-dependant) Github Action runs, considerable amount of time is taken to download the artifact that is all of the Openfire distribution files.

This commit attempts to speed this up by replacing the artifact consisting of many files with an artifact that has only one file: a tar archive of all of those files.