OF-2752: Sessions should be resumable only if closed due to 'unclean' disconnect
Prior to this commit, a session was (stream management) resumable whenever it got closed.

When a session is closed 'cleanly' (eg, with an exchange of `</stream>`), then a session should never be resumed.

This commit attempts to reverse the strategy. It is implied that a session is never resumable, unless it is closed unexpectedly (for example, as caused by a network disconnect).