OF-284: Make ad-hoc 'session data' non-optional
Prior to this, an optional SessionData object was passed along when executing stages of an ad-hoc command. This object was non-null only when the command had multiple stages. The rationale for this was that only multi-staged commands needed to carry over session state.

The session data also includes a bit of metadata (the JID of the actor executing the command). This is desirable to have even when executing only a one-stage command.

This commit makes the sessionData instance a non-optional argument.

(cherry picked from commit 558b4619336fc0073903a32b1e609e4ae6793005)