OF-2755: Fix bug in cluster cleanup code
When a cluster node leaves the cluster, cleanup code is executed. Part of this code is a check that verifies if a certain cache contains entries for each incoming server session. This compares incoming server sessions from another data structure with the content of that cache.

This code previously assumed that there could only be one session per remote domain, which is a false assumption. Moreover, it didn't properly verify the StreamID value.

The assumption in the old code causes an exception to be thrown, which prevents the 'left cluster' event from being finished completely, leading to state inconsistencies.

Although it is likely unexpected that Openfire maintains multiple sessions per remote domain, having more than one is entirely feasible. The code should allow for this.