OF-2774: Fix whitespace ping / keepalive
By placing the whitespace ping handlers beyond the XMPP Decoder, the white space (that needs to be acted on) was not detected (it got normalized away by the XMPP decoder).

This commit places the whitespace ping handlers in front of the XMPP Decoder, so that they can properly detect the whitespace data.