OF-2775: Ensure that RSS dates are parsed in English locale
Prior to this change, the dates from Ignite's RSS feed were parsed in the system locale. When that is not English, the RSS feed would fail to be visible as a result of a parsing error.

(cherry picked from commit a7297320b49f9cc1cad2ac47eaf09147e2c96ccc)