OF-284: Additional commands for XEP-0133 (#2381)
* OF-284: Add Missing Commands for XEP-0133 "Service Administration"

This commit adds almost all ad-hoc commands that are defined in XEP-0133, but were missing in Openfire.

Notable exceptions:
- Get User Password - I deem it unsafe to store/expose plaintext credentials
- Get User Last Login Time - Openfire does not track this data
- Get User Statistics
- Set Message of the Day - MotD functionality could possibly be added in the MotD plugin
- Edit Message of the Day
- Delete Message of the Day
- Set Welcome Message - Welcome message functionality could possibly be added in the Registration plugin
- Delete Welcome Message
- Restart Service - Unsafe: what if the stop works, but the start does not?
- Shut Down Service - Unsafe: we'll end up with admins complaining about an unreachable server

* OF-284: Applied i18n to all adhoc commands

All hardcoded, human readable labels have been replaces with a translatable i18n property value.

* Update xmppserver/src/main/java/org/jivesoftware/openfire/commands/admin/EditAdminList.java

Co-authored-by: Dan Caseley <dan@caseley.me.uk>

* OF-284: Make ad-hoc 'session data' non-optional

Prior to this, an optional SessionData object was passed along when executing stages of an ad-hoc command. This object was non-null only when the command had multiple stages. The rationale for this was that only multi-staged commands needed to carry over session state.

The session data also includes a bit of metadata (the JID of the actor executing the command). This is desirable to have even when executing only a one-stage command.

This commit makes the sessionData instance a non-optional argument.

* OF-284: Ad-Hoc system properties and java.time

Replaces JiveGlobals.getProperty with SystemProperty, and long-based time computation with calls to java.time.

* OF-284: Do not error on missing optional data

* Update xmppserver/src/main/java/org/jivesoftware/openfire/commands/admin/GetListRegisteredUsers.java

Co-authored-by: Dan Caseley <dan@caseley.me.uk>

* Default email to an empty string when missing

* Reduce log level of optional database API mismatch

* OF-284: Fix XML element reuse

An XML element can only have one parent. By re-adding the same element to another parent, an error is thrown. Instead, use a copy that is detached.

* OF-284: Better error messages for invalid JIDs

* OF-284: Switch AuthenticateUser to use jid-single

To be consistent with other fields, and makes sense anyway. Was causing an issue with Spark's registry of fields when 2 fields in the same namespace had different types.

* OF-284: Fix duplicate field value setting

---------

Co-authored-by: Dan Caseley <dan@caseley.me.uk>