OF-2651: Give distinct names to Netty threads
Netty uses various Executor services, that each generate threads. These threads are given a default name, which leads to thread names like: `nioEventLoopGroup-4-1`

This name is not very descriptive. It is helpful, for debugging purposes, to have a name that identifies the source / functionality that relates to the a particular thread.

This commit restores the naming that was used prior to switching to Netty.

(cherry picked from commit 8dff883cf14d50de406dd6202811ff5e237b6e3e)