OF-2792: Admin Console Cache Summary for Clustering
When clustering is used, many caches are not byte-size based, but entity-count based. The admin console page should properly show this.

This commit modifies the Cache interface to provide additional data used to display this page. Specifically:
- Caches now report their capacity unit (byte or entity)
- Caches can report additional remarks (eg: "this is a per clusternode count")

The Cache Overview admin console page has been modified to make use of these new attributes. When combined with an update of the Hazelcast plugin (for it to report the correct values), this improves the correctness of the displayed data remarkably.

The changes in this commit are intended to be backwards compatible with older versions of the Hazelcast plugin. The data reported will still be wrong, but functinality should not break.