OF-2805: Align session details with session summary
When the same user logs in with more than one client, its 'session details' page on the admin console lists a small table with all the sessions for that user.

This table re-uses the session-details row from the session-summary page. This page was modified in OF-2706 (notably, the 'resource' column got deleted).

With this commit, the layout on the session details page is updated to match that of the rows on the session summary page.

(cherry picked from commit 37dfa08326a4c93c1d76520dd882a8f216792164)