OF-2807: Fix event handling for Group Property changes
When a GroupProperty is added, modified or deleted, an event listening mechanism is activated. This, among other things, is used to clean up cache content and push changes to rosters.

This commit fixes a problem with the 'modified' handler. It caused the wrong event to be emitted ('deleted', rather than 'modified').

(cherry picked from commit 09185134c7b45174de06de7e11334884f7c0750c)