OF-2777: Allow groups to be shared with other groups
When a group is to be shared with other groups, this is stored in the database as a group property of name `sharedRoster.showInRoster` having a value of `onlyGroup`. The group property value for `sharedRoster.groupList` controls to what groups this group is shared. It either is a comma-separated list (denoting group names), or it's empty. An empty value signals that the group is shared only with members of the same group.

This commit corrects how the admin panel interprets the database values, and fixes the `getSharedWithUsersInGroupNames` method of Group.

(cherry picked from commit f6a6b39e3a552841a57a87919de363ec614b614d)