OF-2818: Netty should not use 5MB buffers
This corrects a mistake where a configuration of having a _maximum_ buffer size was turned into having a _fixed_ buffer size.

Having 5MB buffers is bad, as with a limited amount of messages, a _lot_ of memory is being consumed.

This commit removes the hard-coded buffer size configuration, while retaining the maximum message length directive.

I've tested that, using a websocket client connection:
- a message of 128k characters can be sent using the default configuration;
- a message of 128k characters can not be sent when Openfire is reconfigured to use 4k for the `xmpp.parser.buffer.size` property