OF-2824: Annotate Session's address to be non-null
Any Session's `address` property is expected to be non-null by the existing implementation. One notable exemption to this is `RemoteIncomingServerSession`, where the address _field_ is lazily initialized (and thus _can_ be null). Its getter value is guaranteed to be non-null though.

This commit annotates and asserts that various Session address fields, getters and setters to have non-null values.

The intention of this change is to simplify implementations (by dropping unneeded null-checks), more strictly document expected state (for easier maintenance) and introduce fail-fast behavior (that should help catch any current or future bugs).

It is not expected that this change will introduce runtime changes (as the values already are non-null).