OF-2824: Annotate ClientSession's address to be a _full_ JID
Any Session's address is non-null, but _client_ sessions also are guarenteed to be full JIDs (in other words: they have a non-null resource part).

This commit explicitly checks for this (to fail fast on any current and future bugs), and removes unneeded checks on usage. This intends to reduce code maintenance costs, by having less complex code.

It is not expected that this change will introduce runtime changes (as the values already are full JIDs).