OF-2824: Relax introduced fail-fast behavior
In commit ef582e131f41bcb61bfb75c083986b92b0bcd002 fail-fast behavior was introduced that causes exceptions to be thrown when certain methods are invoked using a bare JID while the expected argument is a full JID.

For certain lookups, these unexpected bare JIDs appear to be used all the time, causing Openfire to become pretty unusable.

In this commit, the invocations that do not _modify_ cache content are short-circuited when a bare JID argument is used. This has the benefit of preventing the acquisition of a cluster-wide lock.