OF-2832: When merging data, handle null values
When merging VCard data, if either the left-hand or right-hand side data is null, simply return the other bit. The code already did this in one direction, but now does it for the other direction too.

This allows a VCard for a non-existing user to be set (which fixes setting avatars for MUC rooms).