OF-2843 & OF-2844: MUC ban improvements
An admin cannot ban an owner from a MUC room. XEP-0045 section 9.1 specifies:

> [A] user cannot be banned by an admin with a lower affiliation. Therefore, if an admin attempts to ban an owner, the service MUST deny the request and return a `<not-allowed/>` error to the sender

Openfire currently does send an error, but uses the `<conflict/>` condition. Instead, it must use the `<not-allowed/>` condition as specified by the XEP.

XEP-0045 Section 9.1 defines:

> If an admin or owner attempts to ban himself, the service MUST deny the request and return a `<conflict/>` error to the sender. (Note: This is different from the recommended service behavior on kicking oneself.)

Openfire currently allows admins to ban themselves. That should not be possible, as defined by the XEP.