OF-2854: Package of distribution module changed to 'pom'
The distribution module is a metadata module; it does not contain any resources that are directly to be exposed as an artifact (instead, it uses the Assembly and Install4J plugins to generate artifacts).

The module had the package type of 'jar', indicating that a jar artifact was to be produced. This resulted in an empty `distribution.jar` file to be generated.

By switching to the `pom` package type, this no longer is the case.