OF-2857: Replace Install4j runtime library
This replaces the bundled install4j runtime library with one the is obtained from the author's Maven repository.

By no longer bundling the library, a nasty Maven warning that was shown during build disappears.

The replacement library is significantly newer than the original (the old version was never published in a Maven repository).

As a result:
- functional changes might have been introduced
- the Uninstall action that is part of Openfire source code was rewritten

The version that's used now is the same version as the version of Install4j that is used in Bamboo to generate the installers.